Move config_reader to new config package
As part of #422 moving the existing config_reader to the new config
package where config handling and the graph capabilities will eventually
reside.
