+++ integration-tests/sceptre-project/config/7/A.yaml
@@ -1,3 +1,3 @@
 sceptre_user_data:
   type: AWS::CloudFormation::WaitConditionHandle
-template_path: templates/jinja/valid_template.j2
+template_path: templates/jinja/invalid_template_missing_attr.j2
