+++ integration-tests/sceptre-project/config/8/C.yaml
