+++ integration-tests/steps/stacks.py
@@ -8,7 +8,7 @@ from botocore.exceptions import ClientError
 from helpers import read_template_file, get_cloudformation_stack_name
 from helpers import retry_boto_call
 
-from sceptre.config_reader import ConfigReader
+from sceptre.config.reader import ConfigReader
 
 
 def set_stack_timeout(context, stack_name, stack_timeout):
