+++ sceptre/config/__init__.py
@@ -1,0 +1,17 @@
+# -*- coding: utf-8 -*-
+
+import logging
+
+
+__author__ = 'Cloudreach'
+__email__ = 'sceptre@cloudreach.com'
+
+
+# Set up logging to ``/dev/null`` like a library is supposed to.
+# http://docs.python.org/3.3/howto/logging.html#configuring-logging-for-a-library
+class NullHandler(logging.Handler):  # pragma: no cover
+    def emit(self, record):
+        pass
+
+
+logging.getLogger('sceptre').addHandler(NullHandler())
